# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Kilde til informasjon om innhold på et gitt sted.\n[https://punkt.artsdatabanken.no/](https://punkt.artsdatabanken.no/) or on \n[https://artsdatabanken.slack.com/, #åpnedata](https://artsdatabanken.slack.com/).\n",
    "version": "1.0.0",
    "title": "Punkt API",
    "contact": {
      "email": "bjorn.reppen@artsdatabanken.no"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/v1",
  "schemes": [
    "https"
  ],
  "paths": {
    "/punkt": {
      "get": {
        "summary": "Slå opp innhold fra lengde- og breddegrad",
        "description": "",
        "operationId": "getPunkt",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "Lengdegrad i WGS-84 desimalgrader",
            "required": true,
            "default": 63,
            "type": "number"
          },
          {
            "name": "lng",
            "in": "query",
            "description": "Breddegrad i WGS-84 desimalgrader",
            "required": true,
            "default": 10.5,
            "type": "number"
          },
          {
            "name": "layer",
            "in": "query",
            "description": "Only return specified layer, specify the key from the result json",
            "required": false,
            "default": "",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/punktvektor": {
      "get": {
        "summary": "Slå opp innhold i vektoriserte kart fra lengde- og breddegrad",
        "description": "",
        "operationId": "getPunkt",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "Lengdegrad i WGS-84 desimalgrader",
            "required": true,
            "default": 63,
            "type": "number"
          },
          {
            "name": "lng",
            "in": "query",
            "description": "Breddegrad i WGS-84 desimalgrader",
            "required": true,
            "default": 10.5,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/gpx": {
      "get": {
        "summary": "Konverter liste med koordinater til GPX-format.",
        "description": "Brukes for å serve GPX-fil til Kartverket Høydedata API.",
        "operationId": "getGpx",
        "produces": [
          "application/gpx+xml"
        ],
        "parameters": [
          {
            "name": "punkter",
            "in": "query",
            "description": "Kommaseparert liste med lengdegrad,breddegrad,lengdegrad,breddegrad...",
            "required": true,
            "default": "63,10,64,10.5",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/hoydeprofil/diagram": {
      "get": {
        "summary": "Lager høydeprofil visualisering i PNG-format.",
        "description": "Returnerer en json med URLen til bildet.  Se https://www.kartverket.no/data/Hoydeprofil/",
        "operationId": "getHoydeprofilDiagram",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "punkter",
            "in": "query",
            "description": "Kommaseparert liste med lengdegrad,breddegrad,lengdegrad,breddegrad...",
            "required": true,
            "default": "63,10,64,10.5",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/hoydeprofil/json": {
      "get": {
        "summary": "Lager høydeprofil statistikk som JSON.",
        "description": "Se https://www.kartverket.no/data/Hoydeprofil/",
        "operationId": "getHoydeprofilJson",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "punkter",
            "in": "query",
            "description": "Kommaseparert liste med lengdegrad,breddegrad,lengdegrad,breddegrad...",
            "required": true,
            "default": "63,10,64,10.5",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about punkt-oppslag-api on GitHub",
    "url": "https://github.com/Artsdatabanken/punkt-oppslag-api"
  }
}